Projects structure reordanazied 'Render' folder added Color.lua added to work with coloring and pixel-tiles (dependencies upd) Draw.lua added to work with tile drawing (dependencies upd) Checkup.lua added to check on changed pixels Changed pixels stored in global
