+++ scripts/build.lua
@@ -1,3 +1,6 @@
+require('scripts.global')
+require('scripts.render.color')
+
 --- Event handler
 -- Starts update 
 -- @param event Event handlers table
@@ -5,13 +8,13 @@ function on_pixel_built(event)
     --make pixel reference
     local pixel = {
         tile = {
-            name = "pixel-default",
+            name = get_tile_name(Colors.default),
             position = event.created_entity.position
         },
         diode = event.created_entity
     }
 
-    Surface.set_tiles({{position = pixel.diode.position, name = pixel.tile.name}}, true)
+    draw(pixel)
 
     Pixels[pixel.diode.unit_number] = pixel
 end
