+++ scripts/global.lua
@@ -1,43 +1,11 @@
-
 ---Reference to the nauvis surface (only nauvis surface is available for now)
 Surface = nil
 
 ---Array that stores all references to Pixels instances in the game
 Pixels = nil
 
----Table consists of all supported colors for Pixels
-Colors = {
-	["red"] = {
-		r = 1.0, g = .0, b = .0
-	},
-	["green"] = {
-		r = .0, g = 1.0, b = .0
-	},
-	["blue"] = {
-		r = .0, g = .0, b = 1.0
-	},
-	["yellow"] = {
-		r = 1.0, g = 1.0, b = .0
-	},
-	["pink"] = {
-		r = 1.0, g = 0, b = 1.0
-	},
-	["cyan"] = {
-		r = 0, g = 1.0, b = 1.0
-	},
-	["white"] = {
-		r = 1.0, g = 1.0, b = 1.0
-	},
-	["grey"] = {
-		r = .5, g = .5, b = .5
-	},
-	["black"] = {
-		r = .0, g = .0, b = .0
-	},
-	["default"] = {
-	    r = 1, g = 1, b = 1
-	}
-}
+---List of pending pixels to update
+Redraw_Queue = nil
 
 
 --- Generate initial data
@@ -49,10 +17,9 @@ function initialize()
     end
 
     global.current_surface = game.surfaces[1]
+    global.redraw_queue = { tiles = {} }
 
-
-    Surface = global.current_surface
-    Pixels = global.all_pixels[Surface.name]
+    setup_global_variables()
 end
 
 function setup_global_variables()
@@ -62,4 +29,5 @@ function setup_global_variables()
         --create local references to tables stored in the global table
     Surface = global.current_surface
     Pixels = global.all_pixels[Surface.name]
+    Redraw_Queue = global.redraw_queue
 end
